backend/requirements.txt
Flask
requests
google-generativeai
nuclei
gunicorn  # For deployment
flask-limiter  # For rate-limiting
